Move individual methods from/to FeatureExtensions
